server/requirements.txt
fastapi
uvicorn
websocket-client
pyaudiowpatch
